Fixed podiums for good.
Klasa's SettingsGateway feature does not allow multiple keys in a settings item of type "any" when the bot restarts; anytime a reload occurs all entries but the first are removed from memory. Everytime I restarted the bot, therefore, most solves in the database were lost. This was fixed by utilizing an array as opposed to a json object datatype for storing results, and duplicates are prevented via an Array.prototype.filter() check before submitting new times.